backend-core/src/api/deps.py
from typing import Generator
from fastapi import Depends, HTTPException, status
from fastapi.security import OAuth2PasswordBearer
from jose import jwt, JWTError
from sqlalchemy.ext.asyncio import AsyncSession

from src.db.session import AsyncSessionLocal
from src.core.security import settings
from src.models.user import User
from src.schemas.token import TokenPayload

oauth2_scheme = OAuth2PasswordBearer(tokenUrl="/v1/auth/login")

async def get_db() -> Generator:
    async with AsyncSessionLocal() as session:
        yield session

async def get_current_user(
    db: AsyncSession = Depends(get_db),
    token: str = Depends(oauth2_scheme)
) -> User:
    credentials_exception = HTTPException(
        status_code=status.HTTP_401_UNAUTHORIZED,
        detail="Could not validate credentials",
        headers={"WWW-Authenticate": "Bearer"},
    )
    try:
        payload = jwt.decode(token, settings.JWT_SECRET_KEY, algorithms=[settings.ALGORITHM])
        user_id: str = payload.get("sub")
        if user_id is None:
            raise credentials_exception
    except JWTError as e:
        # 💡 토큰 디코딩 실패 이유를 로그에 기록하여 디버깅에 활용
        print(f"DEBUG: JWT Decode Error: {e}")
        raise credentials_exception
    
    # 실제 구현에서는 DB 조회가 필요합니다.
    user = await db.get(User, int(user_id))
    if user is None:
        raise credentials_exception
    return user

# [추가됨] 관리자 권한 확인 함수
def get_current_superuser(
    current_user: User = Depends(get_current_user),
) -> User:
    if not current_user.is_superuser:
        raise HTTPException(
            status_code=status.HTTP_403_FORBIDDEN, detail="The user doesn't have enough privileges"
        )
    return current_user