backend-core/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
sqlalchemy==2.0.35
alembic==1.13.1
psycopg2-binary==2.9.9
asyncpg==0.29.0
pgvector==0.2.4
pydantic==2.6.0
pydantic-settings==2.1.0
python-jose[cryptography]==3.3.0
passlib==1.7.4
python-multipart==0.0.6
requests==2.31.0
httpx==0.26.0
celery==5.3.6
redis==4.6.0
boto3==1.34.14
fastapi-limiter==0.1.5
slowapi==0.1.9
authlib==1.3.0
email-validator==2.1.0.post1
fastapi-mail>=1.4.1
aiofiles==23.1.0